Resolve Comments
Added crypto for using MD5 hashes, could maybe be done with standard lib but unsure. Moved all sync files into `internal/sync`.  Done small refactors to the interactions.dart file
