+++ nyxx_interactions/lib/src/internal/sync/ICommandsSync.dart
