+++ nyxx_interactions/lib/src/internal/sync/ManualCommandSync.dart
